Fingers now work in any hand orientation
I don't know why the relative quaternions between the fingers and
reference IMUs was changing if I changed my hand orientation. That
should not happen. But since it was working fine in a specific hand
orientation, I got the idea of taking the relative quaternions of
everything, reference and fingers, to that orientation and moving
everything to that rotation, basically undoing the rotation in math.
Then, doing the relative stuff on that new reference frame worked.

Also fixed a little bug with the pinky finger.

Now I have 4 finger curls working, yeah!!