Driver customized, two jsons for right and left
Now I know how to customize the data sent from each controller, left and
right, to steam vr. Not only that, I can set its values. They match with
the controller profiles.
Now on steam vr I have what looks like 2 virtual controllers that are
always ON and 2 controllers that correspond to my real trackers and that
can be ON or OFF depending on their real state. I should only have 2,
not 4, must check why.
Firmware and controller profiles have been updated to send only 4 finger
values and discard the buttons that are not implemented on the gloves.
I would LOVE to know how to subscribe to the data that is sent from the
trackers, I have no idea how to do this yet.
Lots of documentation debt now...