Updated device json files for hand sides
This makes the gloves identify themselves as right and left as soon as
they are detected by steamvr. The left hand json was also updated to
mirror the rendermodel and match the expected hand orientation. Position
is a bit off but I will keep that for later.

Also removing the drivers folder that has been deprecated.