Adding auxiliary connectors for mixed inputs
I am still not sure what input types I will have, but I am preparing the
board to be flexible, exposing 8 GPIOs, 4 of which are analog to digital
converters.

I have also started to design the hardware interface with a second board
that can be placed below the main one, to add two rp2040 to the project
if needed.