Organizing FW by hand and usb
Each hand has 2 MCU's, since they are going to be programmed by USB, I
figured I would identify them by the silkscreen markings on the board.
USB2 is more or less done but not tested on the left hand, USB1 is
currently just a copy of USB2 but will be worked on next.