Will send messages from usb2 to usb1 periodically
Until now, messages were sent as soon as the cycle reached the send
part. This is not efficient since it may be sending the same message
many times if there is no data available on the IMUs, and this way I can
make usb1 know when a message starts and ends by looking at inactive
time on the uart. I'll also implement a CRC on the message, I've seen
some incorrect data being parsed and the CRC will just drop those
messages. With these two implementations it should be fixed.